display distance to items based on location in all item list pages

+++ app/src/main/java/group/unimelb/vicmarket/retrofit/RetrofitHelper.java
@@ -150,8 +150,8 @@ public class RetrofitHelper {
     }
 
     public void getNearbyItems(Observer<MainItemListBean> observer,
-                               BigDecimal longitude,
-                               BigDecimal latitude,
+                               double longitude,
+                               double latitude,
                                int maxDistance,
                                int page,
                                int category) {
